refactor: 도메인 모델 통합 및 핸들러 리팩토링
- 도메인 모델을 `internal/model` 패키지로 통합하여 코드 구조 개선
- 로그인 핸들러 및 게시글 관련 핸들러에서 도메인 모델을 `model`로 변경하여 일관성 유지
- 새로운 도메인 모델(Category, Comment, LoginRequest, Post) 추가로 기능 확장
